git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/branches/2.9RC@26310 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/databases/Nek5000/avtNek5000FileFormat.C
@@ -48,7 +48,6 @@
 #endif
 #include <float.h>
 
-
 #include <vtkCellArray.h>
 #include <vtkIntArray.h>
 #include <vtkFloatArray.h>
@@ -95,8 +94,6 @@
 #  define STREQUAL(a,b)              strcasecmp(a,b)
 #endif
 #endif
-
-#include <cstdint>
 
 #include <algorithm>
 #include <fstream>
@@ -1541,7 +1538,7 @@ avtNek5000FileFormat::ReadPoints(int element, int timestep)
     GetDomainSizeAndVarOffset(timestepToUseForMesh, NULL, nFloatsInDomain, 
                               d1, d2, d3);
 
-    int64_t iRealHeaderSize = iHeaderSize + (bParFormat ? aBlocksPerFile[iCurrMeshProc]*sizeof(int) : 0);
+    long iRealHeaderSize = iHeaderSize + (bParFormat ? aBlocksPerFile[iCurrMeshProc]*sizeof(int) : 0);
 
     if (bBinary)
     {
@@ -1552,7 +1549,7 @@ avtNek5000FileFormat::ReadPoints(int element, int timestep)
         if (iPrecision == 4)
         {
             float *tmppts = new float[nPts*iDim];
-            fseek(fdMesh, iRealHeaderSize + (int64_t)nFloatsInDomain*sizeof(float)*element, SEEK_SET);
+            fseek(fdMesh, iRealHeaderSize + (long)nFloatsInDomain*sizeof(float)*element, SEEK_SET);
             size_t res = fread(tmppts, sizeof(float), nPts*iDim, fdMesh); (void) res;
             if (bSwapEndian)
                 ByteSwap32(tmppts, nPts*iDim);
@@ -1577,7 +1574,9 @@ avtNek5000FileFormat::ReadPoints(int element, int timestep)
         else
         {
             double *tmppts = new double[nPts*iDim];
-            fseek(fdMesh, iRealHeaderSize + (int64_t)nFloatsInDomain*sizeof(double)*element, SEEK_SET);
+            fseek(fdMesh, iRealHeaderSize +
+                            (long)nFloatsInDomain*sizeof(double)*element,
+                    SEEK_SET);
             size_t res = fread(tmppts, sizeof(double), nPts*iDim, fdMesh); (void) res;
             if (bSwapEndian)
                 ByteSwap64(tmppts, nPts*iDim);
@@ -1605,9 +1604,9 @@ avtNek5000FileFormat::ReadPoints(int element, int timestep)
         float *pts_tmp = pts;
         for (ii = 0 ; ii < nPts ; ii++)
         {
-            fseek(fdMesh, (int64_t)iAsciiMeshFileStart + 
-                          (int64_t)element*iAsciiMeshFileLineLen*nPts + 
-                          (int64_t)ii*iAsciiMeshFileLineLen, SEEK_SET);
+            fseek(fdMesh, (long)iAsciiMeshFileStart + 
+                          (long)element*iAsciiMeshFileLineLen*nPts + 
+                          (long)ii*iAsciiMeshFileLineLen, SEEK_SET);
             if (iDim == 3)
             {
                 int res = fscanf(fdMesh, " %f %f %f", pts_tmp, pts_tmp+1, pts_tmp+2); (void) res;
@@ -1806,28 +1805,28 @@ avtNek5000FileFormat::ReadVar(int timestate, int element, const char *varname)
     if (bParFormat)
         element = aBlockLocs[element*2 + 1];
 
-    int iRealHeaderSize = iHeaderSize + (bParFormat ? aBlocksPerFile[iCurrVarProc]*sizeof(int) : 0);
+    long iRealHeaderSize = iHeaderSize + (bParFormat ? aBlocksPerFile[iCurrVarProc]*sizeof(int) : 0);
 
     if (bBinary)
     {
-        int64_t filepos;
+        long filepos;
         if (!bParFormat)
-            filepos = (int64_t)iRealHeaderSize + ((int64_t)nFloatsInDomain*element + iBinaryOffset)*sizeof(float);
+            filepos = (long)iRealHeaderSize + ((long)nFloatsInDomain*element + iBinaryOffset)*sizeof(float);
         else
         {
             // This assumes uvw for all fields comes after the mesh as [block0: 216u 216v 216w]...
             // then p or t as   [block0: 216p][block1: 216p][block2: 216p]...
             if (strcmp(varname+2, "velocity") == 0)
             {
-                filepos  = (int64_t)iRealHeaderSize +                              //header
-                           (int64_t)iHasMesh*aBlocksPerFile[iCurrVarProc]*nPts*iDim*iPrecision + //mesh
-                           (int64_t)element*nPts*iDim*iPrecision +                  //start of block
-                           (int64_t)(varname[0] - 'x')*nPts*iPrecision;            //position within block
+                filepos  = (long)iRealHeaderSize +                              //header
+                           (long)iHasMesh*aBlocksPerFile[iCurrVarProc]*nPts*iDim*iPrecision + //mesh
+                           (long)element*nPts*iDim*iPrecision +                  //start of block
+                           (long)(varname[0] - 'x')*nPts*iPrecision;            //position within block
             }
             else
-                filepos = (int64_t)iRealHeaderSize + 
-                          (int64_t)aBlocksPerFile[iCurrVarProc]*iBinaryOffset*iPrecision + //the header, mesh, vel if present,
-                          (int64_t)element*nPts*iPrecision;
+                filepos = (long)iRealHeaderSize + 
+                          (long)aBlocksPerFile[iCurrVarProc]*iBinaryOffset*iPrecision + //the header, mesh, vel if present,
+                          (long)element*nPts*iPrecision;
         }
         if (iPrecision==4)
         {
@@ -1856,10 +1855,10 @@ avtNek5000FileFormat::ReadVar(int timestate, int element, const char *varname)
         float *var_tmp = var;
         for (ii = 0 ; ii < nPts ; ii++)
         {
-            fseek(fdVar, (int64_t)iAsciiCurrFileStart + 
-                         (int64_t)element*iAsciiCurrFileLineLen*nPts + 
-                         (int64_t)ii*iAsciiCurrFileLineLen + 
-                         (int64_t)iAsciiOffset, SEEK_SET);
+            fseek(fdVar, (long)iAsciiCurrFileStart + 
+                         (long)element*iAsciiCurrFileLineLen*nPts + 
+                         (long)ii*iAsciiCurrFileLineLen + 
+                         (long)iAsciiOffset, SEEK_SET);
             int res = fscanf(fdVar, " %f", var_tmp); (void) res;
             var_tmp++;
         }
@@ -2046,18 +2045,18 @@ avtNek5000FileFormat::ReadVelocity(int timestate, int element)
     if (bParFormat)
         element = aBlockLocs[element*2 + 1];
 
-    int iRealHeaderSize = iHeaderSize + (bParFormat ? aBlocksPerFile[iCurrVarProc]*sizeof(int) : 0);
+    long iRealHeaderSize = iHeaderSize + (bParFormat ? aBlocksPerFile[iCurrVarProc]*sizeof(int) : 0);
 
     if (bBinary)
     {
-        int64_t filepos;
+        long filepos;
         if (!bParFormat)
-            filepos = (int64_t)iRealHeaderSize + (int64_t)(nFloatsInDomain*element + iBinaryOffset)*sizeof(float);
+            filepos = (long)iRealHeaderSize + (long)(nFloatsInDomain*element + iBinaryOffset)*sizeof(float);
         else
             //This assumes [block 0: 216u 216v 216w][block 1: 216u 216v 216w]...[block n: 216u 216v 216w]
-            filepos = (int64_t)iRealHeaderSize + 
-                      (int64_t)aBlocksPerFile[iCurrVarProc]*iBinaryOffset*iPrecision + //the header and mesh if one exists
-                      (int64_t)element*nPts*iDim*iPrecision;
+            filepos = (long)iRealHeaderSize + 
+                      (long)aBlocksPerFile[iCurrVarProc]*iBinaryOffset*iPrecision + //the header and mesh if one exists
+                      (long)element*nPts*iDim*iPrecision;
         if (iPrecision == 4)
         {
             float *tmppts = new float[nPts*iDim];
@@ -2117,10 +2116,10 @@ avtNek5000FileFormat::ReadVelocity(int timestate, int element)
         float *var_tmp = var;
         for (ii = 0 ; ii < nPts ; ii++)
         {
-            fseek(fdVar, (int64_t)iAsciiCurrFileStart + 
-                         (int64_t)element*iAsciiCurrFileLineLen*nPts + 
-                         (int64_t)ii*iAsciiCurrFileLineLen + 
-                         (int64_t)iAsciiOffset, SEEK_SET);
+            fseek(fdVar, (long)iAsciiCurrFileStart + 
+                         (long)element*iAsciiCurrFileLineLen*nPts + 
+                         (long)ii*iAsciiCurrFileLineLen + 
+                         (long)iAsciiOffset, SEEK_SET);
             if (iDim == 3)
             {
                 int res = fscanf(fdVar, " %f %f %f", var_tmp, var_tmp+1, var_tmp+2); (void) res;
@@ -2783,11 +2782,11 @@ avtNek5000FileFormat::GetBoundingBoxIntervalTree(int timestep)
 
     for (ii = iRank; ii < iNumOutputDirs; ii+=nProcs)
     {
-        int64_t iFileSizeWithoutMetaData = 136 
+        long iFileSizeWithoutMetaData = 136 
                 + sizeof(int)*aBlocksPerFile[ii] 
-                + ((int64_t)nFloatsPerDomain)*sizeof(float)*((int64_t)aBlocksPerFile[ii]);
+                + ((long)nFloatsPerDomain)*sizeof(float)*((long)aBlocksPerFile[ii]);
 
-        int64_t iMDSize = (nFloatsPerDomain * 2 * sizeof(float) * aBlocksPerFile[ii]) / 
+        long iMDSize = (nFloatsPerDomain * 2 * sizeof(float) * aBlocksPerFile[ii]) / 
                     (iBlockSize[0]*iBlockSize[1]*iBlockSize[2]);
 
         GetFileName(timestep, ii, blockfilename, (int)fileTemplate.size() + 64);
@@ -2798,7 +2797,7 @@ avtNek5000FileFormat::GetBoundingBoxIntervalTree(int timestep)
             break;
         }
         f.seekg( 0, std::ios_base::end );
-        int64_t iFileSize = f.tellg();
+        long iFileSize = f.tellg();
         if (iFileSize < iFileSizeWithoutMetaData+iMDSize)
         {
             errorReadingData = 1;
@@ -3022,13 +3021,13 @@ avtNek5000FileFormat::GetDataExtentsIntervalTree(int timestep, const char *var)
 
     for (ii = iRank; ii < iNumOutputDirs; ii+=nProcs)
     {
-        int64_t iFileSizeWithoutMetaData = 136 
+        long iFileSizeWithoutMetaData = 136 
                 + sizeof(int)*aBlocksPerFile[ii] 
-                + ((int64_t)nFloatsPerDomain)*sizeof(float)*((int64_t)aBlocksPerFile[ii]);
+                + ((long)nFloatsPerDomain)*sizeof(float)*((long)aBlocksPerFile[ii]);
 
-        int64_t iBBSize = 2*iDim * sizeof(float) * aBlocksPerFile[ii];
-        int64_t iDESize = 2 * sizeof(float) * aBlocksPerFile[ii] * numVars;
-        int64_t iMDSize = iBBSize + iDESize;
+        long iBBSize = 2*iDim * sizeof(float) * aBlocksPerFile[ii];
+        long iDESize = 2 * sizeof(float) * aBlocksPerFile[ii] * numVars;
+        long iMDSize = iBBSize + iDESize;
 
         GetFileName(timestep, ii, blockfilename, (int)fileTemplate.size() + 64);
         f.open(blockfilename);
@@ -3038,7 +3037,7 @@ avtNek5000FileFormat::GetDataExtentsIntervalTree(int timestep, const char *var)
             break;
         }
         f.seekg( 0, std::ios_base::end );
-        int64_t iFileSize = f.tellg();
+        long iFileSize = f.tellg();
         if (iFileSize != iFileSizeWithoutMetaData+iMDSize)
         {
             iBBSize = 0;
@@ -3077,7 +3076,7 @@ avtNek5000FileFormat::GetDataExtentsIntervalTree(int timestep, const char *var)
             else
                 varIndex = (bHasVelocity ? iDim : 0) + (bHasPressure ? 1 : 0) + 
                        (bHasTemperature ? 1 : 0) + sComp;
-            int64_t offsetForDE = varIndex*2*sizeof(float)*aBlocksPerFile[ii];
+            long offsetForDE = varIndex*2*sizeof(float)*aBlocksPerFile[ii];
             f.seekg(iFileSizeWithoutMetaData+iBBSize+offsetForDE, std::ios_base::beg);
             f.read( (char *)(bounds + nPrecedingBlocks*2), aBlocksPerFile[ii]*2*sizeof(float) );
         }
